fix(clean-up): removed redeem selectors hook. updated redeem summary card to use selectors. created new selectors instead of infering data in parts of the app
